requirements.txt
# Core
langchain==0.2.16
# openai==1.50.2
# tiktoken==0.7.0
chromadb==0.5.4
pydantic==2.8.2
python-dotenv==1.0.1
langchain-core==0.2.39
langchain-community==0.2.16
langchain-openai==0.1.22
langchain-text-splitters==0.2.2

# Data
pandas==2.2.2
numpy==1.26.4
faker==26.0.0

#Vector Stores
sentence-transformers==2.7.0
torch==2.4.1

# App
streamlit==1.38.0

#LLM
groq==0.13.0
langchain-groq==0.1.4

# Dev & QA (optional)
# pytest==8.3.2
# pytest-cov==5.0.0
# ruff==0.6.8
# black==24.8.0
# isort==5.13.2
# pre-commit==3.8.0

# Optional model backends (comment out if not used)
# groq==0.10.0
# ollama==0.3.1
